pyproject.toml
[project]
name = "multi-modal-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.15",
    "django>=5.2.4",
    "djangorestframework>=3.16.0",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-groq>=0.3.6",
    "langchain-openai>=0.3.28",
    "pip>=25.1.1",
    "python-dotenv>=1.1.1",
    "tiktoken>=0.9.0",
    "unstructured[all-docs]>=0.18.11",
]
